Remove all-#1 rank detection as UserNotFound
Well, this was an interesting surprise to see this here. I have discovered this EO quirk twice, and worked around it twice with pretty much identical code. Now there's the same workaround in both Etternabot and etternaonline-api lol

This instance of the workaround, albeit the OG one, should be removed since it will have a false positive if there is one day a player who is rank 1 in one skillsets